+++ admin/inc/basic.php
@@ -1312,13 +1312,14 @@ function isDebug(){
 }
 
 /**
- * check gs version is Beta
+ * Check if GS Legacy version is a Beta release
  *
- * @since  3.3.0
+ * @since 3.3.0
+ * @since 2025.1 Return boolean value
  * @return boolean true if beta release
  */
 function isBeta(){
-	return strPos(get_site_version(false),"b");
+	return strpos(get_site_version(false), 'b') !== false;
 }
 
 /**
@@ -1427,11 +1428,27 @@ function strip_content($str, $pattern = '/[({]%.*?%[})]/'){
 /**
  * perform transliteration conversion on string
  * @since  3.3.11
- * @param  str $str string to convert
- * @return str      str after transliteration replacement array ran on it
+ * @since 2025.1 Use Transliterator, GSTRANSLITERATIONMODE
+ * @param str $str String to convert
+ * @return str String after transliteration
  */
 function doTransliteration($str){
 	$str = (string) $str;
+	if ($str == '') return '';
+	$mode = getDef('GSTRANSLITERATIONMODE');
+	if (($mode == 1 || $mode == 2) && function_exists('transliterator_create_from_rules')) {
+		$rule = '';
+		if ($mode == 2) {
+			$translit = getTransliteration();
+			if (!empty($translit)) {
+				$rule = implode('; ', array_map(function($k, $v) { return $k . ' > ' . $v; }, array_keys($translit), $translit)) . '; ';
+			}
+		}
+		$transliterator = transliterator_create_from_rules($rule . ':: Any-Latin ; :: Latin-ASCII ;', Transliterator::FORWARD);
+		if ($transliterator !== null) {
+			return transliterator_transliterate($transliterator, $str);
+		}
+	}
 	$translit = getTransliteration();
 	if (!empty($translit)) {
 		$str = str_replace(array_keys($translit), array_values($translit), $str);