+++ admin/install.php
@@ -174,7 +174,13 @@ get_template('header', GSNAME.' &raquo; '. i18n_r('INSTALLATION') );
 				} else{
 					echo '<tr><td>GD Library</td><td><span class="WARNmsg" >'.i18n_r('NOT_INSTALLED').' - '.i18n_r('WARNING') .'</span></td></tr>';
 				}
-				
+
+				if (in_arrayi('intl', $php_modules)) {
+					echo '<tr><td>Intl Extension</td><td><span class="OKmsg">' . i18n_r('INSTALLED') . ' - ' . i18n_r('OK') . '</span></td></tr>';
+				} else {
+					echo '<tr><td>Intl Extension</td><td><span class="WARNmsg">' . i18n_r('NOT_INSTALLED') . ' - ' . i18n_r('WARNING') . '</span></td></tr>';
+				}
+
 				if  (in_arrayi('zip', $php_modules) ) {
 					echo '<tr><td>ZipArchive</td><td><span class="OKmsg" >'.i18n_r('INSTALLED').' - '.i18n_r('OK').'</span></td></tr>';
 				} else{