+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes of GetSimple Legacy CMS will be documented in this file.
 
+## [2025.1] - 2025.01.22
+
+### Added
+
+- Use PHP Intl extension to improve transliteration.
+- Use new constant `GSTRANSLITERATIONMODE` to set transliteration mode with Intl usage. Define the constant in the gsconfig.php file with value: `1` - transliterate to ASCII with a simple Transliterator rule `:: Any-Latin ; :: Latin-ASCII ;`; `2` - transliterate to ASCII with a custom Transliterator rule started with the rule based on the transliteration table defined in a loaded language file. If the constant is not defined or defined with a different value, the legacy transliteration method is used.
+
+### Changed
+
+- The default transliteration table is case-sensitive and extended with symbols from the Belarusian, Ukrainian and Macedonian alphabets.
+- Transliteration table for the Russian language is based on the ГОСТ 7.79—2000 with removed grave accent symbols.
+
+### Fixed
+
+- Change type of the return value of the function `isBeta()` to boolean.
+
 ## [2024.3] - 2024.12.21
 
 ### Added